Task 3. Change file permissions on a hidden file
n this task, you must determine if a hidden file has incorrect permissions and then change the permissions as needed. This action will further remove unauthorized access and strengthen security on the system.

The file .project_x.txt is a hidden file that has been archived and should not be written to by anyone. (The user and group should still be able to read this file.)

1.Check the permissions of the hidden file .project_x.txt and answer the question that follows.

2.Change the permissions of the file .project_x.txt so that both the user and the group can read, but not write to, the file.